Refactored Driver and symbol lookup
* Added a symbol lookup interface and libelf based implementation
* Replaced remnants of Driver with DeviceInfo
* Removed device_keystone.cpp and device_am57.cpp
* Fixed host runtime build errors with BUILD_DSPC

See https://confluence.itg.ti.com/x/iG8p for details (TI Internal)

+++ host/src/core/dsp/memory_provider_interface.h
@@ -50,7 +50,7 @@ public:
 
     virtual void *MapToHostAddressSpace (DSPDevicePtr64 dsp_addr, size_t size,
                                                  bool is_read) const = 0;
-    virtual void  UnmapFromHostAddressSpace (void* host_addr, uint32_t size,
+    virtual void  UnmapFromHostAddressSpace (void* host_addr, size_t size,
                                                 bool is_write) const = 0;
 
     virtual bool   CacheInv  (void *host_addr, std::size_t size) const = 0;
